LPS-142946 Move classes to '.helper', deprecate classes in '.util'
