+++ portal-kernel/src/com/liferay/portal/kernel/display/context/helper/BaseRequestHelper.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.display.context.util;
+package com.liferay.portal.kernel.display.context.helper;
 
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.Layout;
@@ -37,11 +37,8 @@ import javax.portlet.PortletURL;
 import javax.servlet.http.HttpServletRequest;
 
 /**
- * @author     Iván Zaera
- * @deprecated As of Cavanaugh (7.4.x), replaced by {@link
- *             com.liferay.portal.kernel.display.context.helper.BaseRequestHelper}
+ * @author Iván Zaera
  */
-@Deprecated
 public abstract class BaseRequestHelper {
 
 	public BaseRequestHelper(HttpServletRequest httpServletRequest) {
